requirements.txt
openai
instructor
python-dotenv
chess
pydantic
